Copied some Spreadsheet methods to Cell to allow for expressionTree creation
